requirement.txt
django-admin startproject ecom_site 

cd ecom_site

And then, in your __init__.py of your project (ecommerce_site/__init__.py), add:
import pymysql
pymysql.install_as_MySQLdb()


python manage.py startapp 


for data base - CREATE DATABASE ecommerce_db;


import os
MEDIA_ROOT = os.path.join(str(BASE_DIR), 'media')


(env) E:\all_projects\ecompro\ecom_site>python manage.py createsuperuser
Username (leave blank to use 'dell'): admin
Email address: admin@gmail.com
Password: 
Password (again): 
Error: Your passwords didn't match.
Password: 
Password (again):0911
This password is too short. It must contain at least 8 characters.
This password is entirely numeric.
Bypass password validation and create user anyway? [y/N]: y
Superuser created successfully.